+++ polymorphic/admin/childadmin.py
@@ -62,10 +62,6 @@ class PolymorphicChildModelAdmin(admin.ModelAdmin):
         # If the derived class sets the model explicitly, respect that setting.
         kwargs.setdefault('form', self.base_form or self.form)
 
-        # prevent infinite recursion in django 1.6+
-        if not getattr(self, 'declared_fieldsets', None):
-            kwargs.setdefault('fields', None)
-
         return super(PolymorphicChildModelAdmin, self).get_form(request, obj, **kwargs)
 
     def get_model_perms(self, request):
@@ -186,9 +182,8 @@ class PolymorphicChildModelAdmin(admin.ModelAdmin):
     # ---- Extra: improving the form/fieldset default display ----
 
     def get_fieldsets(self, request, obj=None):
-        # If subclass declares fieldsets, this is respected
-        if (hasattr(self, 'declared_fieldsets') and self.declared_fieldsets) \
-           or not self.base_fieldsets:
+        # If subclass declares fieldsets or fields, this is respected
+        if hasattr(self, 'fieldsets') or hasattr(self, 'fields') or not self.base_fieldsets:
             return super(PolymorphicChildModelAdmin, self).get_fieldsets(request, obj)
 
         # Have a reasonable default fieldsets,